Make mediasoup_sys crate mode idiomatic
